requirements.txt
OpenAI
Flask
python-dotenv
line-bot-sdk
